backend/lambdas/python/pyproject.toml
[tool.poetry]
name = "choiflix-backend-python"
version = "0.1.0"
description = ""
authors = ["john-jaihyek-choi <jaihyek.choi@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
python-dotenv = "1.0.1"
mypy-boto3 = "^1.35.33"
boto3 = "1.35.34"
botocore = "1.35.34"
cryptography = "43.0.1"
mypy-boto3-dynamodb = "1.35.24"
mypy-boto3-secretsmanager = "1.35.0"
validators = "0.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
